refactor and clean up test files for twoSum, createHelloWorldFunction, and plusOne
