+++ Solutions/1_Two_Sum/TS/twoSum.ts
@@ -1,5 +1,3 @@
-// En: Given an array of integers nums and an integer target, return indices of the two numbers such that they add up to target.
-// Sv: Givet en array av heltal nums och ett heltal target, returnera index för de två talen så att de summerar till target.
 export function twoSum(nums: number[], target: number): number[] {
   for (let i = 0; i < nums.length - 1; i++) {
     for (let j = i + 1; j < nums.length; j++) {
