+++ Solutions/2667_Create_Hello_World_Function/TS/createHelloWorldFunction.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, it } from 'vitest';
+import { createHelloWorldFunction } from './createHelloWorldFunction';
+
+describe(createHelloWorldFunction, () => {
+  it("Should return 'Hello World' if the function is called", () => {
+    expect(createHelloWorldFunction()).toBe('Hello World');
+  });
+});
