+++ Solutions/2667_Create_Hello_World_Function/TS/createHelloWorldFunction.ts
@@ -1,0 +1,1 @@
+export const createHelloWorldFunction = () => 'Hello World';
