--- Solutions/2667_Create_Hello_World_Function/createHelloWorldFunction.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { createHelloWorldFunction } from "./CreateHelloWorldFunction";
-
-describe(createHelloWorldFunction, () => {
-  it("Should return 'Hello World' if the function is called", () => {
-    expect(createHelloWorldFunction()).toBe("Hello World");
-  });
-});
