--- Solutions/2667_Create_Hello_World_Function/createHelloWorldFunction.ts
@@ -1,6 +1,0 @@
-// En: Write a function createHelloWorld. It should return a new function that always returns "Hello World".
-// Sv: Skriv en funktion createHelloWorld. Den ska returnera en ny funktion som alltid returnerar "Hello World".
-
-export function createHelloWorldFunction(): string {
-  return "Hello World";
-}
