+++ Solutions/66_Plus_One/TS/plusOne.ts
@@ -1,0 +1,11 @@
+export function plusOne(digits: number[]) {
+  for (let i = digits.length - 1; i >= 0; i--) {
+    if (digits[i] < 9) {
+      digits[i]++;
+      return digits;
+    }
+    digits[i] = 0;
+  }
+  digits.unshift(1);
+  return digits;
+}
